Split useConvertInputs to own file
